+++ src/schema/build-context.ts
@@ -10,8 +10,6 @@ import type { ValidatorFn } from "@/typings/ValidatorFn";
 import type { ContainerGetter, ContainerType } from "@/utils/container";
 import { IOCContainer } from "@/utils/container";
 
-export type DateScalarMode = "isoDate" | "timestamp";
-
 export interface ScalarsTypeMap {
   type: Function;
   scalar: GraphQLScalarType;
@@ -20,7 +18,6 @@ export interface ScalarsTypeMap {
 export type ValidateSettings = boolean | ValidatorOptions;
 
 export interface BuildContextOptions {
-  dateScalarMode?: DateScalarMode;
   scalarsMap?: ScalarsTypeMap[];
   /**
    * Indicates if class-validator should be used to auto validate objects injected into params.
@@ -47,8 +44,6 @@ export interface BuildContextOptions {
 }
 
 export abstract class BuildContext {
-  static dateScalarMode: DateScalarMode;
-
   static scalarsMaps: ScalarsTypeMap[];
 
   static validate: ValidateSettings;
@@ -73,10 +68,6 @@ export abstract class BuildContext {
    * Set static fields with current building context data
    */
   static create(options: BuildContextOptions) {
-    if (options.dateScalarMode !== undefined) {
-      this.dateScalarMode = options.dateScalarMode;
-    }
-
     if (options.scalarsMap !== undefined) {
       this.scalarsMaps = options.scalarsMap;
     }
@@ -124,7 +115,6 @@ export abstract class BuildContext {
    * Restore default settings
    */
   static reset() {
-    this.dateScalarMode = "isoDate";
     this.scalarsMaps = [];
     this.validate = false;
     this.validateFn = undefined;